E-Certificate-Writer/requirements.txt
img2pdf==0.3.3
numpy==1.17.2
pandas==0.25.1
Pillow==6.2.0
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
